End of the period work log
Made level creator, rewrote a bit of main and memory, now the first state comes from a file I can write to using levelCreator.

Undo system and Reset system are still not working. Need to make newmemoryController be a memory controller and not just patch the level file onto memorycontroller