Issue has been found.
There is something very wrong with JFrame in Java 8 or whatever I am compiling on. However JFrame displayes transparent images, for some reason, one specific way of saving transparent images is not compatible with this version of Java, but the later versions of Java it works perfectly.

The extra files j/imagedisplay are me testing why images wouldnt work.

When i would display an image, it would have a black background. Sometimes. Some images wont.

This works perfectly on java 17.