Deepcopying.
Its great that I found out that this exists today. Im not entirely sure how I ever got undo or resetlevel working without it before, but I guess that was some magic stroke of luck.

It really, really annoys me that I cant just make a copy of an object in java, and that I have to make a deepcopy instead. IMO, it really should be the other way around, with copy working as, you know, copy, and copyreference copying a reference to the object.

Now that thats fixed, undo and resetlevels magically appear to work out of nowhere with just a few lines of edits, throwing some deepcopies here and there.